refactor(extractor): moving Phind-Chat sources extracting logic
